requirements.txt
Flask
boto3
PyMySQL
python-dotenv
requests
Flask-SQLAlchemy
Flask-Login
cryptography
